rename config.js to index.js
